+++ homework_03/main.py
@@ -1,11 +1,14 @@
 import uvicorn
 from fastapi import FastAPI
-from items_views import router as items_router, prefix as item_prefix
-from users.views import router as users_router, prefix as users_prefix
+
+from view import router as homework_03_router, prefix as homework_03_prefix
+from homework_03.items.items_views import router as items_router, prefix as item_prefix
+from homework_03.users.views import router as users_router, prefix as users_prefix
 
 app = FastAPI()
 app.include_router(items_router, prefix=item_prefix)
 app.include_router(users_router, prefix=users_prefix)
+app.include_router(homework_03_router, prefix=homework_03_prefix)
 
 
 @app.get("/")
